+++ src/pages/News/News.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { Constants } from "../../constant/constant";
+import { Constants } from "../../constant";
 
 const News = () => {
   return <Navigate to={`/news/${Constants.LATEST_NEWS_YEAR}`} />;
