+++ src/pages/admin/Notice/AdminNoticeDetail.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import NoticeForm from "../../../components/admin/form/NoticeForm";
-import { Constants } from "../../../constant/constant";
+import { Constants } from "../../../constant";
 import Title from "../../../layouts/Title";
 import { useNotices } from "../../../recoills/notices";
